Moved navbar so that it is now under dashboard routes and not under the root layout. Did some organizing of the ui components folders and files as a result of the navbar change
