+++ app/dashboard/page.tsx
@@ -1,5 +1,3 @@
-
-
 export default function Page() {
     return <div>The dashboard page</div>
 }
